# Conflicts:
#	appveyor.yml
#	src/NLog.Extended/LayoutRenderers/AppSettingLayoutRenderer.cs
#	src/NLog/Common/InternalLoggerMessageEventArgs.cs
#	src/NLog/Config/LoggingConfigurationParser.cs
#	src/NLog/Config/XmlLoggingConfiguration.cs
#	src/NLog/GDC.cs
#	src/NLog/Internal/FileCharacteristics.cs
#	src/NLog/Internal/FileInfoHelper.cs
#	src/NLog/Internal/Files/FileCharacteristics.cs
#	src/NLog/Internal/Files/FileCharacteristicsHelper.cs
#	src/NLog/Internal/Files/PortableFileCharacteristicsHelper.cs
#	src/NLog/Internal/Files/Win32FileCharacteristicsHelper.cs
#	src/NLog/Internal/PortableFileCharacteristicsHelper.cs
#	src/NLog/Internal/StringHelpers.cs
#	src/NLog/Internal/Win32FileCharacteristicsHelper.cs
#	src/NLog/LayoutRenderers/EventContextLayoutRenderer.cs
#	src/NLog/LayoutRenderers/FuncThreadAgnosticLayoutRenderer.cs
#	src/NLog/LayoutRenderers/Log4JXmlEventLayoutRenderer.cs
#	src/NLog/LogReceiverService/ILogReceiverClient.cs
#	src/NLog/LogReceiverService/SoapLogReceiverClient.cs
#	src/NLog/LogReceiverService/WcfILogReceiverClient.cs
#	src/NLog/MDC.cs
#	src/NLog/NDC.cs
#	src/NLog/Targets/IJsonSerializer.cs
#	tests/NLog.UnitTests/Config/RuleConfigurationTests.cs
#	tests/NLog.UnitTests/LayoutRenderers/ExceptionTests.cs
#	tests/NLog.UnitTests/NLog.UnitTests.csproj

+++ src/NLog/Targets/FileTarget.cs
@@ -1,5 +1,5 @@
 // 
-// Copyright (c) 2004-2019 Jaroslaw Kowalski <jaak@jkowalski.net>, Kim Christensen, Julian Verdurmen
+// Copyright (c) 2004-2020 Jaroslaw Kowalski <jaak@jkowalski.net>, Kim Christensen, Julian Verdurmen
 // 
 // All rights reserved.
 // 
@@ -89,11 +89,11 @@ namespace NLog.Targets
         /// <summary>
         /// List of the associated file appenders with the <see cref="FileTarget"/> instance.
         /// </summary>
-        private FileAppenderCache _fileAppenderCache;
+        private IFileAppenderCache _fileAppenderCache;
 
         IFileArchiveMode GetFileArchiveHelper(string archiveFilePattern)
         {
-            return _fileArchiveHelper ?? (_fileArchiveHelper = FileArchiveModeFactory.CreateArchiveStyle(archiveFilePattern, ArchiveNumbering, GetArchiveDateFormatString(ArchiveDateFormat), ArchiveFileName != null, MaxArchiveFiles));
+            return _fileArchiveHelper ?? (_fileArchiveHelper = FileArchiveModeFactory.CreateArchiveStyle(archiveFilePattern, ArchiveNumbering, GetArchiveDateFormatString(ArchiveDateFormat), ArchiveFileName != null, MaxArchiveFiles > 0 || MaxArchiveDays > 0));
         }
         private IFileArchiveMode _fileArchiveHelper;
 
@@ -108,6 +108,11 @@ namespace NLog.Targets
         /// The maximum number of archive files that should be kept.
         /// </summary>
         private int _maxArchiveFiles;
+
+        /// <summary>
+        /// The maximum days of archive files that should be kept.
+        /// </summary>
+        private int _maxArchiveDays;
 
         /// <summary>
         /// The filename as target
@@ -146,13 +151,20 @@ namespace NLog.Targets
         /// <remarks>
         /// The default value of the layout is: <code>${longdate}|${level:uppercase=true}|${logger}|${message:withexception=true}</code>
         /// </remarks>
-        public FileTarget()
+        public FileTarget() : this(FileAppenderCache.Empty)
+        {
+        }
+
+        internal FileTarget(IFileAppenderCache fileAppenderCache)
         {
             ArchiveNumbering = ArchiveNumberingMode.Sequence;
             _maxArchiveFiles = 0;
+            _maxArchiveDays = 0;
+
             ConcurrentWriteAttemptDelay = 1;
             ArchiveEvery = FileArchivePeriod.None;
             ArchiveAboveSize = ArchiveAboveSizeDisabled;
+            ArchiveOldFileOnStartupAboveSize = 0;
             ConcurrentWriteAttempts = 10;
             ConcurrentWrites = true;
             Encoding = Encoding.UTF8;
@@ -169,7 +181,7 @@ namespace NLog.Targets
             ForceManaged = false;
             ArchiveDateFormat = string.Empty;
 
-            _fileAppenderCache = FileAppenderCache.Empty;
+            _fileAppenderCache = fileAppenderCache;
             CleanupFileName = true;
 
             WriteFooterOnArchivingOnly = false;
@@ -481,6 +493,8 @@ namespace NLog.Targets
         [Advanced]
         public int ConcurrentWriteAttemptDelay { get; set; }
 
+        private bool? _archiveOldFileOnStartup;
+
         /// <summary>
         /// Gets or sets a value indicating whether to archive old log file on startup.
         /// </summary>
@@ -490,7 +504,23 @@ namespace NLog.Targets
         /// </remarks>
         /// <docgen category='Output Options' order='10' />
         [DefaultValue(false)]
-        public bool ArchiveOldFileOnStartup { get; set; }
+        public bool ArchiveOldFileOnStartup
+        {
+            get => _archiveOldFileOnStartup ?? false;
+            set => _archiveOldFileOnStartup = value;
+        }
+
+        /// <summary>
+        /// Gets or sets a value of the file size threshold to archive old log file on startup.
+        /// </summary>
+        /// <remarks>
+        /// This option won't work if <see cref="ArchiveOldFileOnStartup"/> is set to <c>false</c>
+        /// Default value is 0 which means that the file is archived as soon as archival on
+        /// startup is enabled.
+        /// </remarks>
+        /// <docgen category='Output Options' order='10' />
+        [DefaultValue(0)]
+        public long ArchiveOldFileOnStartupAboveSize { get; set; }
 
         /// <summary>
         /// Gets or sets a value specifying the date format to use when archiving files.
@@ -532,7 +562,7 @@ namespace NLog.Targets
             get => _archiveAboveSize;
             set
             {
-                if ((_archiveAboveSize == ArchiveAboveSizeDisabled) != (value == ArchiveAboveSizeDisabled))
+                if ((ArchiveAboveSize > ArchiveAboveSizeDisabled) != (value > ArchiveAboveSizeDisabled))
                 {
                     _archiveAboveSize = value;
                     ResetFileAppenders("ArchiveAboveSize Changed"); // Reset archive file-monitoring
@@ -629,6 +659,24 @@ namespace NLog.Targets
                 {
                     _maxArchiveFiles = value;
                     ResetFileAppenders("MaxArchiveFiles Changed");  // Enforce archive cleanup
+                }
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the maximum days of archive files that should be kept.
+        /// </summary>
+        /// <docgen category='Archival Options' order='10' />
+        [DefaultValue(0)]
+        public int MaxArchiveDays
+        {
+            get => _maxArchiveDays;
+            set
+            {
+                if (_maxArchiveDays != value)
+                {
+                    _maxArchiveDays = value;
+                    ResetFileAppenders("MaxArchiveDays Changed");  // Enforce archive cleanup
                 }
             }
         }
@@ -879,7 +927,7 @@ namespace NLog.Targets
                 return SingleProcessFileAppender.TheFactory;
         }
 
-        private bool IsArchivingEnabled => ArchiveAboveSize != ArchiveAboveSizeDisabled || ArchiveEvery != FileArchivePeriod.None;
+        private bool IsArchivingEnabled => ArchiveAboveSize > ArchiveAboveSizeDisabled || ArchiveEvery != FileArchivePeriod.None;
 
         private bool IsSimpleKeepFileOpen => KeepFileOpen && !NetworkWrites && !ReplaceFileContentsOnEachWrite && !ConcurrentWrites;
 
@@ -1624,25 +1672,28 @@ namespace NLog.Targets
                 FinalizeFile(fileName, isArchiving: true);
             }
 
-            if (string.Equals(Path.GetDirectoryName(archiveFileName.FileName), fileInfo.DirectoryName, StringComparison.OrdinalIgnoreCase))
+            if (fileArchiveStyle.IsArchiveCleanupEnabled)
             {
-                // Extra handling when archive-directory is the same as logging-directory
-                for (int i = 0; i < existingArchiveFiles.Count; ++i)
+                if (string.Equals(Path.GetDirectoryName(archiveFileName.FileName), fileInfo.DirectoryName, StringComparison.OrdinalIgnoreCase))
                 {
-                    if (string.Equals(existingArchiveFiles[i].FileName, fileInfo.FullName, StringComparison.OrdinalIgnoreCase))
+                    // Extra handling when archive-directory is the same as logging-directory
+                    for (int i = 0; i < existingArchiveFiles.Count; ++i)
                     {
-                        existingArchiveFiles.RemoveAt(i);
-                        break;
+                        if (string.Equals(existingArchiveFiles[i].FileName, fileInfo.FullName, StringComparison.OrdinalIgnoreCase))
+                        {
+                            existingArchiveFiles.RemoveAt(i);
+                            break;
+                        }
                     }
                 }
-            }
 
-            existingArchiveFiles.Add(archiveFileName);
+                existingArchiveFiles.Add(archiveFileName);
 
-            var cleanupArchiveFiles = fileArchiveStyle.CheckArchiveCleanup(archiveFilePattern, existingArchiveFiles, MaxArchiveFiles);
-            foreach (var oldArchiveFile in cleanupArchiveFiles)
-            {
-                DeleteOldArchiveFile(oldArchiveFile.FileName);
+                var cleanupArchiveFiles = fileArchiveStyle.CheckArchiveCleanup(archiveFilePattern, existingArchiveFiles, MaxArchiveFiles, MaxArchiveDays);
+                foreach (var oldArchiveFile in cleanupArchiveFiles)
+                {
+                    DeleteOldArchiveFile(oldArchiveFile.FileName);
+                }
             }
 
             return archiveFileName.FileName;
@@ -1809,8 +1860,8 @@ namespace NLog.Targets
         /// <returns>Filename to archive. If <c>null</c>, then nothing to archive.</returns>
         private string GetArchiveFileName(string fileName, LogEventInfo ev, int upcomingWriteSize, DateTime previousLogEventTimestamp)
         {
-            var hasFileName = !(fileName == null && _previousLogFileName == null);
-            if (hasFileName)
+            fileName = fileName ?? _previousLogFileName;
+            if (!string.IsNullOrEmpty(fileName))
             {
                 return GetArchiveFileNameBasedOnFileSize(fileName, upcomingWriteSize) ??
                        GetArchiveFileNameBasedOnTime(fileName, ev, previousLogEventTimestamp);
@@ -1841,10 +1892,7 @@ namespace NLog.Targets
                 return _previousLogFileName;
             }
 
-            //this is an expensive call
-            var fileLength = _fileAppenderCache.GetFileLength(fileName);
-            string fileToArchive = fileLength != null ? fileName : _previousLogFileName;
-            return fileToArchive;
+            return fileName;
         }
 
         /// <summary>
@@ -1855,21 +1903,30 @@ namespace NLog.Targets
         /// <returns>Filename to archive. If <c>null</c>, then nothing to archive.</returns>
         private string GetArchiveFileNameBasedOnFileSize(string fileName, int upcomingWriteSize)
         {
-            if (ArchiveAboveSize == ArchiveAboveSizeDisabled)
+            if (ArchiveAboveSize <= ArchiveAboveSizeDisabled)
             {
                 return null;
             }
 
             var previousFileName = GetPotentialFileForArchiving(fileName);
-            if (previousFileName == null)
+            if (string.IsNullOrEmpty(previousFileName))
             {
                 return null;
             }
 
-            var length = _fileAppenderCache.GetFileLength(previousFileName);
-            if (length == null)
+            //this is an expensive call
+            var fileLength = _fileAppenderCache.GetFileLength(previousFileName);
+            if (!fileLength.HasValue)
             {
-                return null;
+                if (!string.IsNullOrEmpty(_previousLogFileName) && previousFileName != _previousLogFileName)
+                {
+                    upcomingWriteSize = 0;
+                    return GetArchiveFileNameBasedOnFileSize(_previousLogFileName, upcomingWriteSize);
+                }
+                else
+                {
+                    return null;
+                }
             }
 
             if (previousFileName != fileName)
@@ -1877,13 +1934,18 @@ namespace NLog.Targets
                 upcomingWriteSize = 0;  // Not going to write to this file
             }
 
-            var shouldArchive = length.Value + upcomingWriteSize > ArchiveAboveSize;
+            var shouldArchive = ShouldArchiveOnFileSize(fileLength.Value, upcomingWriteSize);
             if (shouldArchive)
             {
-                return previousFileName;
+                return previousFileName;    // Will re-check if archive is still necessary after flush/close file
             }
 
             return null;
+        }
+
+        private bool ShouldArchiveOnFileSize(long fileLength, int upcomingWriteSize)
+        {
+            return fileLength + upcomingWriteSize > ArchiveAboveSize;
         }
 
         /// <summary>
@@ -1901,16 +1963,54 @@ namespace NLog.Targets
             }
 
             fileName = GetPotentialFileForArchiving(fileName);
-            if (fileName == null)
+            if (string.IsNullOrEmpty(fileName))
             {
                 return null;
             }
 
+            DateTime? creationTimeSource = TryGetArchiveFileCreationTimeSource(fileName, previousLogEventTimestamp);
+            if (!creationTimeSource.HasValue)
+            {
+                if (!string.IsNullOrEmpty(_previousLogFileName) && fileName != _previousLogFileName)
+                {
+                    return GetArchiveFileNameBasedOnTime(_previousLogFileName, logEvent, previousLogEventTimestamp);
+                }
+                else
+                {
+                    return null;
+                }
+            }
+
+            DateTime fileCreateTime = TruncateArchiveTime(creationTimeSource.Value, ArchiveEvery);
+            DateTime logEventTime = TruncateArchiveTime(logEvent.TimeStamp, ArchiveEvery);
+            if (fileCreateTime != logEventTime)
+            {
+                string formatString = GetArchiveDateFormatString(string.Empty);
+                string fileCreated = creationTimeSource.Value.ToString(formatString, CultureInfo.InvariantCulture);
+                string logEventRecorded = logEvent.TimeStamp.ToString(formatString, CultureInfo.InvariantCulture);
+                var shouldArchive = fileCreated != logEventRecorded;
+                if (shouldArchive)
+                {
+                    return fileName;    // Will re-check if archive is still necessary after flush/close file
+                }
+            }
+
+            return null;
+        }
+
+        private DateTime? TryGetArchiveFileCreationTimeSource(string fileName, DateTime previousLogEventTimestamp)
+        {
             // Linux FileSystems doesn't always have file-birth-time, so NLog tries to provide a little help
             DateTime? fallbackTimeSourceLinux = (previousLogEventTimestamp != DateTime.MinValue && IsSimpleKeepFileOpen) ? previousLogEventTimestamp : (DateTime?)null;
             var creationTimeSource = _fileAppenderCache.GetFileCreationTimeSource(fileName, fallbackTimeSourceLinux);
-            if (creationTimeSource == null)
+            if (!creationTimeSource.HasValue)
+                return null;
+
+            var fileLength = _fileAppenderCache.GetFileLength(fileName);   // Verifies file-handle by checking FileStream.Length
+            if (!fileLength.HasValue)
             {
+                InternalLogger.Debug("FileTarget(Name={0}): Cannot get length of file {1} with creation date {2}.", Name, fileName, creationTimeSource.Value);
+                _initializedFiles.Remove(fileName);
                 return null;
             }
 
@@ -1930,22 +2030,7 @@ namespace NLog.Targets
                 }
             }
 
-            DateTime fileCreateTime = TruncateArchiveTime(creationTimeSource.Value, ArchiveEvery);
-            DateTime logEventTime = TruncateArchiveTime(logEvent.TimeStamp, ArchiveEvery);
-            if (fileCreateTime != logEventTime)
-            {
-                string formatString = GetArchiveDateFormatString(string.Empty);
-                string fileCreated = creationTimeSource.Value.ToString(formatString, CultureInfo.InvariantCulture);
-                string logEventRecorded = logEvent.TimeStamp.ToString(formatString, CultureInfo.InvariantCulture);
-
-                var shouldArchive = fileCreated != logEventRecorded;
-                if (shouldArchive)
-                {
-                    return fileName;
-                }
-            }
-
-            return null;
+            return creationTimeSource;
         }
 
         /// <summary>
@@ -2184,6 +2269,32 @@ namespace NLog.Targets
         }
 
         /// <summary>
+        /// Decision logic whether to archive logfile on startup.
+        /// <see cref="ArchiveOldFileOnStartup"/> and <see cref="ArchiveOldFileOnStartupAboveSize"/> properties.
+        /// </summary>
+        /// <param name="fileName">File name to be written.</param>
+        /// <returns>Decision whether to archive or not.</returns>
+        internal bool ShouldArchiveOldFileOnStartup(string fileName)
+        {
+            if (_archiveOldFileOnStartup == false)
+            {
+                // explicitly disabled and not the default
+                return false;
+            }
+            
+            var aboveSizeSet = ArchiveOldFileOnStartupAboveSize > 0;
+            if (aboveSizeSet)
+            {
+                // Check whether size threshold exceeded
+                var length = _fileAppenderCache.GetFileLength(fileName);
+                return length.HasValue && length.Value > ArchiveOldFileOnStartupAboveSize;
+            }
+
+            // No size threshold specified, use archiveOldFileOnStartup flag
+            return _archiveOldFileOnStartup == true;
+        }
+
+        /// <summary>
         /// Invokes the archiving and clean up of older archive file based on the values of
         /// <see cref="NLog.Targets.FileTarget.ArchiveOldFileOnStartup"/> and
         /// <see cref="NLog.Targets.FileTarget.DeleteOldFileOnStartup"/> properties respectively.
@@ -2195,20 +2306,20 @@ namespace NLog.Targets
             InternalLogger.Debug("FileTarget(Name={0}): Process file '{1}' on startup", Name, fileName);
             RefreshArchiveFilePatternToWatch(fileName, logEvent);
 
-            if (ArchiveOldFileOnStartup)
+            try
             {
-                try
+                if (ShouldArchiveOldFileOnStartup(fileName))
                 {
                     DoAutoArchive(fileName, logEvent, DateTime.MinValue, true);
                 }
-                catch (Exception exception)
-                {
-                    InternalLogger.Warn(exception, "FileTarget(Name={0}): Unable to archive old log file '{1}'.", Name, fileName);
+            }
+            catch (Exception exception)
+            {
+                InternalLogger.Warn(exception, "FileTarget(Name={0}): Unable to archive old log file '{1}'.", Name, fileName);
 
-                    if (exception.MustBeRethrown())
-                    {
-                        throw;
-                    }
+                if (exception.MustBeRethrown())
+                {
+                    throw;
                 }
             }
 
@@ -2218,13 +2329,13 @@ namespace NLog.Targets
             }
 
             string archiveFilePattern = GetArchiveFileNamePattern(fileName, logEvent);
-            if (!string.IsNullOrEmpty(archiveFilePattern) && FileArchiveModeFactory.ShouldDeleteOldArchives(MaxArchiveFiles))
+            if (!string.IsNullOrEmpty(archiveFilePattern))
             {
                 var fileArchiveStyle = GetFileArchiveHelper(archiveFilePattern);
-                if (fileArchiveStyle.AttemptCleanupOnInitializeFile(archiveFilePattern, MaxArchiveFiles))
+                if (fileArchiveStyle.AttemptCleanupOnInitializeFile(archiveFilePattern, MaxArchiveFiles, MaxArchiveDays))
                 {
                     var existingArchiveFiles = fileArchiveStyle.GetExistingArchiveFiles(archiveFilePattern);
-                    var cleanupArchiveFiles = fileArchiveStyle.CheckArchiveCleanup(archiveFilePattern, existingArchiveFiles, MaxArchiveFiles);
+                    var cleanupArchiveFiles = fileArchiveStyle.CheckArchiveCleanup(archiveFilePattern, existingArchiveFiles, MaxArchiveFiles, MaxArchiveDays);
                     foreach (var oldFile in cleanupArchiveFiles)
                     {
                         DeleteOldArchiveFile(oldFile.FileName);
